Replace special characters for PTX.
On Windows, the differences in from Itanium to MS mangling mean that the PTX backend trips over the MS mangled names, which are used in device code in SSCP.
Therefore, we replace all unsupported characters (`@`, `<`, `>`, `? ` mostly), with `_` in device code.
I just *hope* that this won't lead to new Symbol clashes.
Note, it will obviously destroy the demanglability of the names - potentially we could write our own demangling tool and use special encodings for what we replace, to make the original name recoverable.. but not sure how important/useful this will be.